Split api atomic txs related into separated file
